Reorganize src dir
